msm: camera: Proper probe deferral in pproc driver
Error path in cpp/vpe driver was not correct in case of unavailable clock.
This caused a kernel crash. This patch adds proper handling of error
during cpp initialization:
In case of error during probe, the allocated resources are released and
returned error is propagated.
The patch adds similar handling in error during open.

Change-Id: I910688d7d567f3db3d44ed815db2576f6df7258b
Signed-off-by: Petar Sivenov <psiven@codeaurora.org>