iommu: msm: create some helper macros for generating register addresses
We already have some macros that can be used to read and write Iommu
registers. The register address calculations are built into the macros
themselves. In order to facilitate register address calculations in
other parts of the code, refactor the current register accessor macros
so that the address calculations are done in separate, reusable macros.

Change-Id: Id1eb4cb1f5ef7239188329b3d4f025ee62dc7b37
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>