usb: ehci-msm: Fix host mode HUB resume issue during runtime suspend
When the micro a cable is connected over hub, hub is auto suspended
and usb entered into low power mode. If usb device is connected to
the suspended hub then root hub is resumed, it doesn't handle the
device resume for received USB asynchronous interrupt and the connected
devices are not resumed.

To fix this drive resume signaling in host mode to wakeup peripherals

CRs-Fixed: 747418
Change-Id: I49899cdd86aa04ecde5e34ff07da8202921ff7c3
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>