pwm: qpnp: do not control EN_PWM_OUTPUT for LPG lite peripherals
In LPG lite peripherals, the EN_PWM_OUTPUT bit controls whether
or not the LPG_OUT signal is output onto DTEST as opposed to
whether or not the LPG_OUT signal is output in general. DTEST
output is not desired so ensure that the EN_PWM_OUTPUT bit is
never set for LPG lite peripherals.

CRs-Fixed: 803896
Change-Id: I34b7832e2b15e60b5bf2d9bcc0837af67f6d5400
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>