perf: Tighten (and fix) the grouping condition
The fix from 9fc81d87420d ('perf: Fix events installation during
moving group') was incomplete in that it failed to recognise that
creating a group with events for different CPUs is semantically
broken -- they cannot be co-scheduled.

Furthermore, it leads to real breakage where, when we create an event
for CPU Y and then migrate it to form a group on CPU X, the code gets
confused where the counter is programmed -- triggered by the fuzzer.

Fix this by tightening the rules for creating groups. Only allow
grouping of counters that can be co-scheduled in the same context.
This means for the same task and/or the same cpu.

Fixes: 9fc81d87420d ('perf: Fix events installation during moving group')
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>